Add files via upload
This file is the workaround to a certain problem we were having with the UART0 receiver module (more details further down). This file works with the Integrated Glove Final file. 

The problem with the UART0 receiver module that we were trying to solve with this workaround is the following:

The receiver array does not synchronize with the transmitter array. That is, the transmitter MCU sends information regarding the flex sensors to the receiver MCU in the form of an array. The transmitter MCU has a transmit array, and the receiver MCU has a receive array. Logically, the two arrays should be identical. However, it was observed that the data in the receive array are not perfectly aligned with that in the transmit array. 

For instance, suppose that in the transmit array, we have the values 0x1F, 0x2E, 0x3D, and 0x4C in indices 1,2,3 and 4. However, in the receive array, these values appear in indices, say, 3,4,5 and 6. Furthermore, these values then change their position in the receive array --  they shift to, say, 2,3,4, and 5. Then they shift to 4,5,6, and 7, and so on. The receive array values are not in sync with the transmit array values.

The issue was that the receiver MCU needed to poll the UART0 rx pin perfectly in sync with the transmitter MCU UART0 tx pin. If the rx pin was not polled at the correct time, the receiver would lose data (or the data would arrive late in the next iteration) and hence the wrong data would fill the wrong index. The receiver MCU needed to know exactly when to poll so that the correct data will be received. 

This file attempts to fix this issue by using UART1 as the module that synchronizes the polling done by UART0. Whenever UART1 receives any character, an interrupt is generated and this signals UART0 to start polling.